Have update_ocsp generate OCSPs in temp location
OCSP files are no longer written directly to the target directory (which would
possibly overwrite a good OCSP with an empty file if there was a failure), but
instead are written to a temporary location first. If the file is not empty and
the openssl command did not return a non-zero exit status, only then is the
target file path updated.