Readability and reliability fixes
Ensures we don't fail on "servers" entries when pillar returns None
instead of a list (such as the case may be when a backend is not
running, such as app-legacy on staging). We should probably do this
everywhere, but that would result in a lot of very long lines for
something unnecessary in practice.

Improves loops by iterating over key/value pairs instead of a tuple
in a couple of places.